core/serializers/products.py
from rest_framework import serializers

from core.models import Products


class ProductsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'

class ProductsCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'


class ProductsUpdateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'


class ProductsDeleteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = ['id']
