pyproject.toml
[tool.poetry]
name = "test-work-1"
version = "0.1.0"
description = ""
authors = ["Andrew-715 <kutuzkin98@mail.ru>"]
readme = "README.md"
packages = [{include = "test_work_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
